refactor(modList): updateInfo を main プロセスへ移設し tRPC 化
list.json のダウンロードと dataURL.packages の組み立てを main の
modList サービスに移し、renderer 側 updateInfo は tRPC 呼び出しの
薄い委譲にする。呼び出し元(core/package/convertId/setting)は
変更不要。Settings タブの React 化で UI が tRPC だけで完結する
ための前提整理。
